+++ tools/ds_lint_migrator/lib/ds_lint_migrator.dart
@@ -1,7 +1,9 @@
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 import 'src/rules/avoid_ds_alert_old_severity.dart';
+import 'src/rules/avoid_ds_badge_old_color_api.dart';
 import 'src/rules/avoid_ds_button_old_label_param.dart';
+import 'src/rules/avoid_ds_text_field_old_hint_param.dart';
 
 PluginBase createPlugin() => _DsLintMigratorPlugin();
 
@@ -10,5 +12,7 @@ class _DsLintMigratorPlugin extends PluginBase {
   List<LintRule> getLintRules(CustomLintConfigs configs) => [
         AvoidDsButtonOldLabelParam(),
         AvoidDsAlertOldSeverity(),
+        AvoidDsTextFieldOldHintParam(),
+        AvoidDsBadgeOldColorApi(),
       ];
 }